+++ src/gen_rules.ml
@@ -192,10 +192,14 @@ module Gen(P : sig val sctx : Super_context.t end) = struct
           let open Sexp.Encoder in
           constr "exclude" (List.map ~f:Path.to_sexp js_targets)
         ) in
-        Predicate.create ~id ~f:(fun p ->
-          not (List.exists js_targets ~f:(Path.equal p)))
+        List.iter js_targets ~f:(fun js_target ->
+          assert (Path.equal (Path.parent_exn js_target) ctx_dir));
+        Predicate.create ~id ~f:(fun basename ->
+          not (List.exists js_targets ~f:(fun js_target ->
+            String.equal (Path.basename js_target) basename)))
       in
-      Build.paths_matching ~dir:ctx_dir ~loc:Loc.none pred
+      File_selector.create ~dir:ctx_dir pred
+      |> Build.paths_matching ~loc:Loc.none
     in
     Build_system.Alias.add_deps
       ~dyn_deps