+++ src/odoc.ml
@@ -322,11 +322,11 @@ let odocs =
   let odoc_pred =
     Glob.of_string_exn (Loc.of_pos __POS__) "*.odoc"
     |> Glob.to_pred
-    |> Predicate.contramap ~f:Path.basename ~map_id:Fn.id
   in
   fun ctx target ->
     let dir = Paths.odocs ctx target in
-    Build_system.eval_pred ~dir odoc_pred
+    File_selector.create ~dir odoc_pred
+    |> Build_system.eval_pred
     |> Path.Set.fold ~init:[] ~f:(fun d acc ->
       create_odoc ctx d ~target :: acc)