+++ src/variant.ml
@@ -12,3 +12,7 @@ let mt_posix   = make "mt_posix"
 let byte       = make "byte"
 let native     = make "native"
 let plugin     = make "plugin"
+
+let encode t = Dune_lang.atom_or_quoted_string (to_string t)
+
+let decode = Dune_lang.Decoder.plain_string (fun ~loc:_ s -> make s)