+++ test/blackbox-tests/test-cases/variants/multiple-implementations-for-variants/lib2.default/vlib.ml
@@ -1,0 +1,1 @@
+let run () = Printf.printf "hello from lib2.default"