+++ test/blackbox-tests/test-cases/variants/resolution-priority/direct.ocaml/direct.ml
@@ -1,0 +1,1 @@
+let run () = Printf.printf "hi from direct.ocaml\n"