+++ test/blackbox-tests/test-cases/vlib-default-impl/dependency-cycle/clock.c/clock.ml
@@ -1,0 +1,1 @@
+let run () = Printf.printf "hi from clock.c"