+++ CHANGES.md
@@ -11,6 +11,12 @@ unreleased
 
 - Fix glob dependencies on installed directories (#1965, @rgrinberg)
 
+- Add support for library variants and default implementations. (#1900,
+  @TheLortex)
+
+- Add experimental `$ dune init` command. This command is used to create or
+  update project boilerplate. (#1448, fixes #159, @shonfeder)
+
 1.8.2 (10/03/2019)
 ------------------